pyproject.toml
[tool.poetry]
name = "twumber"
version = "0.1.0"
description = "Discord bot that reacts to a twitter link with a number of images"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.8.5"
"discord.py" = "^2.0.1"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
